AVD: unify tab, add Spark view, Dockerized Spark, and Docker helpers
Overview:
- Split AVD tab: moved legacy dashboard into app/tabs/avd_classic.py and
  made app/tabs/avd.py a thin orchestrator with two sub-tabs (Classic + Spark).
- Added generic Spark tab (app/tabs/spark_generic.py) to auto-discover
  and render <prefix>_spark_* collections.

Spark:
- Rewrote spark/streaming_job.py as a generic job with modes:
  * top_artists (default), top_tracks, feature_avg.
- Robust normalization for artist fields and feature casting.
- Mongo sink via foreachBatch with richer metadata (prefix/mode/feature/group).
- New image + service: spark/Dockerfile and docker-compose service "spark-app".

App:
- team_config.yaml: added explicit tab order (Overview, avd, avd_spark)
  and labeled the AVD Spark tab.

Tooling:
- makefile: big cleanup and new targets for host vs docker workflows:
  run/run-avd, run-docker, consume-host/consume-docker, app-host/app-docker,
  spark-build/up/logs/down, env-host/env-spark, show-env, test-env.
- Producers now respect container envs (dotenv override disabled in Docker).

Compose:
- docker-compose.yml: added spark-app service with correct in-network endpoints.

Notes:
- Keep .env files out of git; use envs/.env.host și envs/.env.spark pentru preseturi.
