+++ src/app/team_config.yaml
@@ -1,18 +1,12 @@
+ui:
+  tabs_order:
+    - Overview
+    - avd
+    - avd_spark
+
 team:
   - prefix: avd
-    display_name: "Avram Dorin-Emilian"
-  - prefix: gilian
-    display_name: "Rehm Gilian"
-  - prefix: vadim
-    display_name: "Steshkov Vadim"
-  - prefix: alex
-    display_name: "Manzl Alexander"
-  - prefix: raw
-    display_name: Generic RAW
+    display_name: Dorin-Emilian Avram
+  - prefix: avd_spark
+    display_name: AVD · Spark
 
-defaults:
-  mongo_url: "mongodb://root:example@localhost:27017/?authSource=admin"
-  db_name: "spotify_db"
-
-ui:
-  tabs_order: ["Overview", "dorin", "gilian", "vadim", "alex", "Generic RAW"]
